feat: store bignumber as bytes32[] instead of bytes (#369)
* feat: add BigNumber gas tests

* feat: store bignumber as bytes32[] instead of bytes

This commit changes the data model for bignumber to use bytes32[]
instead of bytes, so we can more easily encode, decode, and iterate over
data word-by-word instead of byte-by-byte. Gas savings are huge for
`from` and `asBytes` but slightly worse for some of the math operations.
Hoping to fix those in coming PRs

* feat: add BigNumber gas tests

* fix: parallel tests

* feat: fix tests

we were adding random padding in the middle of the bytes array, this pr
moves it to the front so we properly encode/decode

* fix: lint

* fix: coverage

* fix: review comments

- 32 -> 0x20
- remove unused vars
- update snapshots

* fix: comments

* prettier

* snapshot updated

* Eco 5 minor issues (#370)

* feat: add BigNumber gas tests

* feat: store bignumber as bytes32[] instead of bytes

This commit changes the data model for bignumber to use bytes32[]
instead of bytes, so we can more easily encode, decode, and iterate over
data word-by-word instead of byte-by-byte. Gas savings are huge for
`from` and `asBytes` but slightly worse for some of the math operations.
Hoping to fix those in coming PRs

* fix: test suite

* feat: store bignumber as bytes32[] instead of bytes

This commit changes the data model for bignumber to use bytes32[]
instead of bytes, so we can more easily encode, decode, and iterate over
data word-by-word instead of byte-by-byte. Gas savings are huge for
`from` and `asBytes` but slightly worse for some of the math operations.
Hoping to fix those in coming PRs

* gas(CVF-37): use bitwise and instead of modulus

x % 2^k is the same as x 2^k-1, and saves a bit of gas

* fix(CVF-42/43): remove duplicate array length mloads

* fix(CVF-44): remove redundant check

* fix(CVF-46): gt instead of iszero(eq(x, 0))

no gas changes, but more readable

* fix(CVF-48): optimize carry calculation

* fix(CVF-53): use shr/shl instead of mul to a power of 2

saves a bit of gas

* fix: formatting

* fix(CVF-62): remove loop for checking empty slots

This saves a whopping 1 gas

* Eco 4 bn optimizations 2 (#392)

* feat: Improve carry calculation

CVF-51

* feat: improve carry math again

ABDK's solution doesn't work exactly as given, but a modified version
here that improves gas a bit

CVF-52

* fix: improve more carry math

CVF-54

* fix: more carry math

CVF-55

* fix: use innerDiff instead of absdiff

absDiff just ensures the first is greater than thes econd, since we know
that diffSquared is greater than res we can use the inner method to save
a bit of checking gas

* fix: use stack variable instead of mem

CVF-58

* Update dependency eslint-plugin-mocha to v10.1.0

* removed backed up package-lock
removed exclusive test

* updated snapshots

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Albert Brown <albertbrown@Bobingy.local>

Co-authored-by: Albert Brown <albertbrown@Bobingy.local>
Co-authored-by: Albert Brown <albert@eco.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>