Circuit breaker (#390)
* feat: add pausable to erc20

* feat: add template circuit breaker election proposal

This commit adds a proposal that would elect a given address to be the
circuit breaker admin. This admin currently has the ability to pause and
unpause ECO and ECOx. As a downstream effect, if ECO is paused, the
proposal fee is waived as it's impossible to collect when ECO is paused.
Potentially we might need to add support for canceling monetary policy
voting as well

Issue: ECO-268

* fix: review comments

* fix: review comments